refactor: :recycle: Move validation to the `PlaygroundOptions` component
